requirements-test.txt
pytest==6.1.1
pytest-django==4.1.0
pytest-cov==2.10.1
pytest-sugar==0.9.4
